+++ src/design-system/tokens/index.ts
@@ -1,6 +1,12 @@
 import tokenSource from './source.json'
 
-export { designTokenCssVariables, designTokenValues, designTokens, designTokensByCssVariable } from './generated'
-export type { DesignTokenCssVariable, DesignTokenPath } from './generated'
+export {
+  designTokenCssVariables,
+  designTokenResolver,
+  designTokenValues,
+  designTokens,
+  designTokensByCssVariable,
+} from './generated/tokens'
+export type { DesignTokenCssVariable, DesignTokenPath } from './generated/tokens'
 
 export const dtcgTokenSource = tokenSource